backendapi/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
requests
pillow
pyodbc
faster-whisper
ffmpeg-python
torch
opencc-python-reimplemented
